#  Emulate metastore calls in BenchmarkInformationSchema