Before the commit:
```
# Run complete. Total time: 00:03:10

Benchmark                                                (queryId)  (schemasCount)  (tablesCount)  Mode  Cnt     Score     Error  Units
BenchmarkInformationSchema.queryInformationSchema        FULL_SCAN             200            200  avgt   10  2863.929 ± 329.840  ms/op
BenchmarkInformationSchema.queryInformationSchema   LIKE_PREDICATE             200            200  avgt   10    26.482 ±   7.837  ms/op
BenchmarkInformationSchema.queryInformationSchema  MIXED_PREDICATE             200            200  avgt   10    61.268 ±  13.492  ms/op
```

After:
```
# Run complete. Total time: 00:15:20

Benchmark                                                (queryId)  (schemasCount)  (tablesCount)  Mode  Cnt      Score     Error  Units
BenchmarkInformationSchema.queryInformationSchema        FULL_SCAN             200            200  avgt   10  55259.057 ± 712.193  ms/op
BenchmarkInformationSchema.queryInformationSchema   LIKE_PREDICATE             200            200  avgt   10     26.804 ±   4.957  ms/op
BenchmarkInformationSchema.queryInformationSchema  MIXED_PREDICATE             200            200  avgt   10    345.492 ±  25.485  ms/op
```

cc: @kokosing